+++ password-instances/ChangeLog.md
@@ -2,10 +2,11 @@
 
 ## 3.0.0.0
 
--   Small update to the README and the `password-instances` library now depends
-    on the new `password-types` package, instead of on the `password` package.
+-   The `password-instances` library now depends on the new
+    `password-types` package, instead of on the `password` package.
     [#40](https://github.com/cdepillabout/password/pull/40)
     Thanks to [@Vlix](https://github.com/Vlix)
+-   Small update to the README
 -   Changed `PersistField` instance of `PasswordHash` to use
     `decodeUtf8'` for better handling of UTF8 decoding exceptions.
     [#43](https://github.com/cdepillabout/password/pull/43)